refactor(ferric): extract fromWireToNative into shared wire module
`worker.ts` and `pool-worker.ts` both carried a copy-pasted helper
that rehydrated wire-format `FerricSymbol` tagged objects into the
native class. The two implementations were structurally identical
and only differed in which `FerricSymbol` constructor they closed
over.

Move the helper into `wire.ts` with the constructor injected as a
parameter, then reduce each worker to a one-line shim that captures
its own native class. This removes the duplication and keeps the
rehydration logic in one place.

Addresses PR #74 review (P2 #5).
