+++ packages/ferric/src/worker.ts
@@ -20,7 +20,7 @@
 import { parentPort } from "node:worker_threads";
 import { resolve } from "node:path";
 import type { WorkerRequest, WorkerResponse, WorkerInit } from "./wire";
-import { ABORT_FLAG_INDEX, RUN_BATCH_SIZE, toWire, isWireSymbol, extractFerricError } from "./wire";
+import { ABORT_FLAG_INDEX, RUN_BATCH_SIZE, toWire, fromWireToNative, extractFerricError } from "./wire";
 import type { NativeEngine } from "./native";
 
 if (!parentPort) {
@@ -50,29 +50,8 @@ const NativeEngine = native["Engine"] as any;
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const NativeFerricSymbol = native["FerricSymbol"] as any;
 
-/**
- * Recursively convert wire symbols in args back to native FerricSymbol
- * before passing to the engine.
- */
-function fromWireToNative(val: unknown): unknown {
-  if (val === null || val === undefined) return val;
-  if (typeof val !== "object") return val;
-
-  if (isWireSymbol(val)) {
-    return new NativeFerricSymbol(val.value);
-  }
-
-  if (Array.isArray(val)) {
-    return val.map(fromWireToNative);
-  }
-
-  // Plain object — convert values recursively (e.g. template slots).
-  const result: Record<string, unknown> = {};
-  for (const [k, v] of Object.entries(val)) {
-    result[k] = fromWireToNative(v);
-  }
-  return result;
-}
+/** Shim the shared helper with this worker's native FerricSymbol constructor. */
+const wireToNative = (val: unknown): unknown => fromWireToNative(val, NativeFerricSymbol);
 
 // ---------------------------------------------------------------------------
 // Engine state
@@ -182,10 +161,10 @@ function handleMethod(method: string, args: unknown[]): unknown {
       return engine.assertString(args[0] as string);
     case "assertFact": {
       const [relation, ...fields] = args as [string, ...unknown[]];
-      return engine.assertFact(relation, ...fields.map(fromWireToNative));
+      return engine.assertFact(relation, ...fields.map(wireToNative));
     }
     case "assertTemplate":
-      return engine.assertTemplate(args[0] as string, fromWireToNative(args[1]) as Record<string, unknown>);
+      return engine.assertTemplate(args[0] as string, wireToNative(args[1]) as Record<string, unknown>);
     case "retract":
       engine.retract(args[0] as number);
       return undefined;
